Fix sql statements for history views
The different view for the history graph zoom levels produced rows
where the timestamp was an integer and such where it was a float.
This lead to warnings when the values were read out. An explicit
cast fixes this.

Also, there is no need to subtract the current time when binning
values. It actually has the unwanted effect, that the bin boundaries
change as time progresses.